+++ src/main/java/com/springboot/clienteapp/models/repository/PerfilRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.springboot.clienteapp.models.entity.Perfil;
 
-public interface PerfilRepository extends CrudRepository<Perfil, Long> {
+public interface PerfilRepository extends CrudRepository<Perfil, Integer> {
 
 }
